RasPi/requirements.txt
azure-iot-device
python-dotenv
RPi.bme280
rich
Flask
